cast
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@126857 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: bc90034ea19017dd283579d46879d69aed0e2a7b

+++ Online/Online/Writer/src/MDFWriterNet.cpp
@@ -679,7 +679,7 @@ StatusCode MDFWriterNet::writeBuffer(void *const /*fd*/, const void *data, size_
 
   static int nbLate=0;
   unsigned int runNumber = getRunNumber(mHeader, len);
-  if (runNumber == -1) { 
+  if ((int)runNumber == -1) { 
       *m_log << MSG::FATAL << WHERE 
              << "Event with runnumber == -1 received. Not forwarding."
              << endmsg;
